feat: 주문 처리 로직 추가 및 주문 관련 엔티티, DTO, 서비스 업데이트

+++ src/main/java/com/shop/shoppingapi/controller/dto/mypage/MyRoomResponse.java
@@ -12,6 +12,6 @@ public class MyRoomResponse {
     private String username;
     private String role;
     private Boolean loggedIn;
-    private BigDecimal point;
+    private int point;
 
 }
